requirements.txt
altair==4.2.0
argon2-cffi @ file:///D:/bld/argon2-cffi_1605217264330/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.24.2
azure-cosmos==4.3.0
azure-identity==1.10.0
azure-keyvault-secrets==4.4.0
azure-storage-blob==12.12.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1612213472466/work
blinker==1.4
boto3==1.21.13
botocore==1.24.13
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi @ file:///C:/ci/cffi_1606237165371/work
chardet @ file:///C:/ci/chardet_1605296577955/work
click==7.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
commonmark==0.9.1
conda==4.13.0
conda-package-handling @ file:///C:/ci/conda-package-handling_1605484860147/work
cryptography @ file:///C:/ci/cryptography_1607618721363/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dominate==2.6.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.1
fonttools==4.25.0
gitdb==4.0.9
GitPython==3.1.27
greenlet @ file:///C:/ci/greenlet_1611958475989/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///D:/bld/importlib-metadata_1622488480664/work
instagram-scraper==1.10.6
ipykernel @ file:///D:/bld/ipykernel_1620913104713/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1622679406477/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
isodate==0.6.1
itsdangerous==1.1.0
jedi @ file:///D:/bld/jedi_1610147001629/work
Jinja2==2.11.2
jmespath==0.10.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter @ file:///C:/ci/jupyter_1607685287094/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-console @ file:///opt/conda/conda-bld/jupyter_console_1647002188872/work
jupyter-core @ file:///D:/bld/jupyter_core_1612125412016/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282555033/work
MarkupSafe @ file:///C:/ci/markupsafe_1607027406824/work
matplotlib @ file:///C:/ci/matplotlib-suite_1638289827560/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
menuinst==1.4.16
mistune @ file:///D:/bld/mistune_1610113097974/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
munkres==1.1.4
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1614336084111/work
nbconvert @ file:///D:/bld/nbconvert_1605402074138/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1621259862661/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1634088717532/work
oauthlib==3.2.0
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas==1.4.3
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
portalocker==2.4.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
pyarrow==8.0.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.7.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1620245170812/work
PyJWT==2.4.0
Pympler==1.0.1
PyMySQL==1.0.2
pyodbc==4.0.32
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
pyparsing==2.4.7
pyrsistent @ file:///D:/bld/pyrsistent_1610147004555/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==228
pywinpty @ file:///C:/ci/pywinpty_1607419945780/work
pyzmq @ file:///D:/bld/pyzmq_1622038686207/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1649078897110/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1649073884068/work
requests @ file:///tmp/build/80754af9/requests_1606691187061/work
requests-oauthlib==1.3.1
rich==12.4.4
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1605524207026/work
s3transfer==0.5.2
semver==2.13.0
Send2Trash==1.5.0
sip==4.19.13
six @ file:///C:/ci/six_1605205404068/work
smmap==5.0.0
SQLAlchemy @ file:///C:/ci/sqlalchemy_1618090016696/work
streamlit==1.10.0
terminado==0.9.4
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
toml==0.10.2
toolz==0.11.2
tornado @ file:///D:/bld/tornado_1610094910414/work
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
validators==0.20.0
visitor==0.1.3
watchdog==2.1.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension @ file:///C:/ci/widgetsnbextension_1644991377168/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
WTForms==3.0.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
pyodbc
sqlalchemy
jinja2